Rename ContentDimensionLoader to ContentDimensionRepository
