+++ Content/Application/MessageHandler/LoadContentViewMessageHandler.php
@@ -13,10 +13,10 @@ declare(strict_types=1);
 
 namespace Sulu\Bundle\ContentBundle\Content\Application\MessageHandler;
 
-use Sulu\Bundle\ContentBundle\Content\Application\ContentDimensionLoader\ContentDimensionLoaderInterface;
 use Sulu\Bundle\ContentBundle\Content\Application\Message\LoadContentViewMessage;
 use Sulu\Bundle\ContentBundle\Content\Domain\Factory\ViewFactoryInterface;
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\ContentViewInterface;
+use Sulu\Bundle\ContentBundle\Content\Domain\Repository\ContentDimensionRepositoryInterface;
 use Sulu\Bundle\ContentBundle\Content\Domain\Repository\DimensionRepositoryInterface;
 
 class LoadContentViewMessageHandler
@@ -27,9 +27,9 @@ class LoadContentViewMessageHandler
     private $dimensionRepository;
 
     /**
-     * @var ContentDimensionLoaderInterface
+     * @var ContentDimensionRepositoryInterface
      */
-    private $contentDimensionLoader;
+    private $contentDimensionRepository;
 
     /**
      * @var ViewFactoryInterface
@@ -38,11 +38,11 @@ class LoadContentViewMessageHandler
 
     public function __construct(
         DimensionRepositoryInterface $dimensionRepository,
-        ContentDimensionLoaderInterface $contentDimensionLoader,
+        ContentDimensionRepositoryInterface $contentDimensionRepository,
         ViewFactoryInterface $viewFactory
     ) {
         $this->dimensionRepository = $dimensionRepository;
-        $this->contentDimensionLoader = $contentDimensionLoader;
+        $this->contentDimensionRepository = $contentDimensionRepository;
         $this->viewFactory = $viewFactory;
     }
 
@@ -50,7 +50,7 @@ class LoadContentViewMessageHandler
     {
         $content = $message->getContent();
         $dimensionCollection = $this->dimensionRepository->findByAttributes($message->getDimensionAttributes());
-        $contentDimensionCollection = $this->contentDimensionLoader->load($content, $dimensionCollection);
+        $contentDimensionCollection = $this->contentDimensionRepository->load($content, $dimensionCollection);
 
         return $this->viewFactory->create($contentDimensionCollection);
     }
