+++ Resources/config/handlers.xml
@@ -17,14 +17,14 @@
 
         <service id="sulu_content.load_content_dimension_handler" class="Sulu\Bundle\ContentBundle\Content\Application\MessageHandler\LoadContentMessageHandler">
             <argument type="service" id="sulu.repository.dimension"/>
-            <argument type="service" id="sulu_content.content_dimension_loader"/>
+            <argument type="service" id="sulu_content.content_dimension_repository"/>
             <argument type="service" id="sulu_content.view_factory"/>
             <argument type="service" id="sulu_content.api_view_resolver"/>
         </service>
 
         <service id="sulu_content.load_content_view_dimension_handler" class="Sulu\Bundle\ContentBundle\Content\Application\MessageHandler\LoadContentViewMessageHandler">
             <argument type="service" id="sulu.repository.dimension"/>
-            <argument type="service" id="sulu_content.content_dimension_loader"/>
+            <argument type="service" id="sulu_content.content_dimension_repository"/>
             <argument type="service" id="sulu_content.view_factory"/>
         </service>
     </services>
