+++ Resources/config/services.xml
@@ -46,18 +46,18 @@
 
         <!-- Content Dimension Factory -->
         <service id="sulu_content.content_dimension_factory" class="Sulu\Bundle\ContentBundle\Content\Application\ContentDimensionFactory\ContentDimensionCollectionFactory">
-            <argument type="service" id="sulu_content.content_dimension_loader"/>
+            <argument type="service" id="sulu_content.content_dimension_repository"/>
             <argument type="tagged" tag="sulu_content.content_dimension_mapper"/>
         </service>
 
         <service id="Sulu\Bundle\ContentBundle\Content\Domain\Factory\ContentDimensionCollectionFactoryInterface" alias="sulu_content.content_dimension_factory"/>
 
         <!-- Content Dimension Loader -->
-        <service id="sulu_content.content_dimension_loader" class="Sulu\Bundle\ContentBundle\Content\Infrastructure\Doctrine\ContentDimensionLoader">
+        <service id="sulu_content.content_dimension_repository" class="Sulu\Bundle\ContentBundle\Content\Infrastructure\Doctrine\ContentDimensionRepository">
             <argument type="service" id="doctrine.orm.entity_manager"/>
         </service>
 
-        <service id="Sulu\Bundle\ContentBundle\Content\Application\ContentDimensionLoader\ContentDimensionLoaderInterface" alias="sulu_content.content_dimension_loader"/>
+        <service id="Sulu\Bundle\ContentBundle\Content\Domain\Repository\ContentDimensionRepositoryInterface" alias="sulu_content.content_dimension_repository"/>
 
         <!-- Route Defaults Provider -->
         <service id="sulu_content.route_defaults_provider" class="Sulu\Bundle\ContentBundle\Content\Infrastructure\Sulu\Route\ContentRouteDefaultsProvider">
