+++ Tests/Unit/Content/Application/MessageHandler/LoadContentViewMessageHandlerTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Sulu\Bundle\ContentBundle\Tests\Unit\Content\Application\MessageHandler;
 
 use PHPUnit\Framework\TestCase;
-use Sulu\Bundle\ContentBundle\Content\Application\ContentDimensionLoader\ContentDimensionLoaderInterface;
 use Sulu\Bundle\ContentBundle\Content\Application\Message\LoadContentViewMessage;
 use Sulu\Bundle\ContentBundle\Content\Application\MessageHandler\LoadContentViewMessageHandler;
 use Sulu\Bundle\ContentBundle\Content\Domain\Factory\ViewFactoryInterface;
@@ -26,18 +25,19 @@ use Sulu\Bundle\ContentBundle\Content\Domain\Model\ContentViewInterface;
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\Dimension;
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\DimensionCollection;
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\DimensionInterface;
+use Sulu\Bundle\ContentBundle\Content\Domain\Repository\ContentDimensionRepositoryInterface;
 use Sulu\Bundle\ContentBundle\Content\Domain\Repository\DimensionRepositoryInterface;
 
 class LoadContentViewMessageHandlerTest extends TestCase
 {
     protected function createLoadContentViewMessageHandlerInstance(
         DimensionRepositoryInterface $dimensionRepository,
-        ContentDimensionLoaderInterface $contentDimensionLoader,
+        ContentDimensionRepositoryInterface $contentDimensionRepository,
         ViewFactoryInterface $viewFactory
     ): LoadContentViewMessageHandler {
         return new LoadContentViewMessageHandler(
             $dimensionRepository,
-            $contentDimensionLoader,
+            $contentDimensionRepository,
             $viewFactory
         );
     }
@@ -94,15 +94,15 @@ class LoadContentViewMessageHandlerTest extends TestCase
             $contentDimension2->reveal(),
         ]);
 
-        $contentDimensionLoader = $this->prophesize(ContentDimensionLoaderInterface::class);
-        $contentDimensionLoader->load($content->reveal(), $dimensionCollection)->willReturn($contentDimensionCollection);
+        $contentDimensionRepository = $this->prophesize(ContentDimensionRepositoryInterface::class);
+        $contentDimensionRepository->load($content->reveal(), $dimensionCollection)->willReturn($contentDimensionCollection);
         $contentView = $this->prophesize(ContentViewInterface::class);
         $viewFactory = $this->prophesize(ViewFactoryInterface::class);
         $viewFactory->create($contentDimensionCollection)->willReturn($contentView->reveal())->shouldBeCalled();
 
         $createContentMessageHandler = $this->createLoadContentViewMessageHandlerInstance(
             $dimensionRepository->reveal(),
-            $contentDimensionLoader->reveal(),
+            $contentDimensionRepository->reveal(),
             $viewFactory->reveal()
         );
